--- src/main/java/rmblworx/tools/timey/CountdownExpiredEvent.java
@@ -1,12 +1,0 @@
-/**
- * 
- */
-package rmblworx.tools.timey;
-
-/**
- * @author mmatthies
- *
- */
-public class CountdownExpiredEvent implements TimeyEvent {
-
-}
