+++ src/main/java/rmblworx/tools/timey/ITimey.java
@@ -1,5 +1,6 @@
 package rmblworx.tools.timey;
 
+import rmblworx.tools.timey.event.TimeyEventListener;
 import rmblworx.tools.timey.exception.EmptyArgumentException;
 import rmblworx.tools.timey.exception.NullArgumentException;
