--- src/main/java/rmblworx/tools/timey/TimeyEvent.java
@@ -1,11 +1,0 @@
-/**
- * 
- */
-package rmblworx.tools.timey;
-
-/**
- * @author mmatthies
- *
- */
-public interface TimeyEvent {
-}
