--- src/main/java/rmblworx/tools/timey/TimeyEventListener.java
@@ -1,13 +1,0 @@
-/**
- * 
- */
-package rmblworx.tools.timey;
-
-/**
- * @author mmatthies
- *
- */
-public interface TimeyEventListener {
-
-	void handleEvent(TimeyEvent timeyEvent);
-}
