+++ src/main/java/rmblworx/tools/timey/TimeyFacade.java
@@ -8,6 +8,8 @@ import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 
+import rmblworx.tools.timey.event.TimeyEventDispatcher;
+import rmblworx.tools.timey.event.TimeyEventListener;
 import rmblworx.tools.timey.exception.EmptyArgumentException;
 import rmblworx.tools.timey.exception.NullArgumentException;
 import rmblworx.tools.timey.vo.TimeDescriptor;
