+++ src/main/java/rmblworx/tools/timey/event/CountdownExpiredEvent.java
@@ -1,0 +1,10 @@
+package rmblworx.tools.timey.event;
+
+/**
+ * Informiert den Listener ueber den Ablauf eines Countdowns.
+ * 
+ * @author mmatthies
+ */
+public class CountdownExpiredEvent implements TimeyEvent {
+
+}
