+++ src/main/java/rmblworx/tools/timey/event/TimeyEvent.java
@@ -1,0 +1,9 @@
+package rmblworx.tools.timey.event;
+
+/**
+ * Marker-Interface zur Kennzeichnung von Implementierungen die als timey-Event fungieren sollen.
+ * 
+ * @author mmatthies
+ */
+public interface TimeyEvent {
+}
