+++ src/main/java/rmblworx/tools/timey/event/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * Enthaelt alle Implementierungen fuer das timey-Eventkonzept.
+ * @author mmatthies
+ */
+package rmblworx.tools.timey.event;
